framework/backend/python/packages/k1s0-observability/pyproject.toml
[project]
name = "k1s0-observability"
version = "0.1.0"
description = "OpenTelemetry-based observability for k1s0 Python services"
requires-python = ">=3.12"
dependencies = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "k1s0-error",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/k1s0_observability"]
